+++ src/components/Field/FieldTypes.tsx
@@ -6,7 +6,7 @@ import { FieldDefColor } from "@root/components/Field/FormFieldColorPicker/Color
 import { FieldDefDate } from "@root/components/Field/FormFieldDate/DateField";
 import { FieldDefDropdown } from "@root/components/Field/FormFieldDropdownSingleSelection";
 import { FieldDefImageUpload } from "@root/components/Field/FormFieldImageUpload";
-import { FieldDefImageVideoDocumentLink } from "@root/forms/FormFieldImageVideoLinkDocumentBrowsing";
+import { FieldDefImageVideoDocumentLink } from "@root/components/Field/FormFieldImageVideoLinkDocumentBrowsing";
 import { FieldDefMapCoordinates } from "@root/forms/FormFieldMapCoordinates";
 import { FieldDefMatrix } from "@root/forms/FormFieldMatrix";
 import { FieldDefNumberTable } from "@root/forms/FormFieldNumberTable";
